requirements.txt
behave==1.2.6
selenium==4.15.2
PyYAML~=6.0.1
allure-behave==2.13.2